requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
fastapi==0.109.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
promise==2.3
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.36.3
starlette-graphene3==0.6.0
strawberry-graphql==0.219.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
